+++ mas-connecta/src/main/java/com/ca/mas/connecta/client/MASConnectaManager.java
@@ -43,7 +43,6 @@ import static com.ca.mas.foundation.MAS.TAG;
 public class MASConnectaManager implements MASConnectaClient, Observer {
 
     private static MASConnectaManager instance = new MASConnectaManager();
-    private long timeOutInMillis = -1;
     private MASConnectOptions connectOptions;
     private MASConnectaListener connectaListener;
     private String clientId;
@@ -63,14 +62,6 @@ public class MASConnectaManager implements MASConnectaClient, Observer {
 
     public void setConnectaListener(MASConnectaListener listener) {
         connectaListener = listener;
-    }
-
-    /**
-     * @deprecated Use {@link MASConnectaManager#disconnect(MASCallback)}
-     */
-    @Deprecated
-    public void stop() {
-        disconnect(null);
     }
 
     private MqttConnecta getMqttConnecta() {
@@ -93,11 +84,6 @@ public class MASConnectaManager implements MASConnectaClient, Observer {
 
         if (connectOptions == null) {
             connectOptions = new MASConnectOptions();
-        }
-        if (timeOutInMillis > 0 && timeOutInMillis >= ConnectaConsts.TIMEOUT_VAL) {
-            // timeout converted to seconds. The default for MQTT is 30 seconds. 0 means no timeout
-            int toSeconds = (int) (timeOutInMillis / ConnectaConsts.SEC_MILLIS);
-            connectOptions.setConnectionTimeout(toSeconds);
         }
 
         //Initialize the credentials and connection Factory for the connect option.
@@ -337,18 +323,6 @@ public class MASConnectaManager implements MASConnectaClient, Observer {
     @Override
     public void setConnectOptions(MASConnectOptions connectOptions) {
         this.connectOptions = connectOptions;
-    }
-
-
-    @Override
-    public void setTimeOutInMillis(long timeOutInMillis) {
-        this.timeOutInMillis = timeOutInMillis;
-    }
-
-
-    @Override
-    public long getTimeOutInMillis() {
-        return timeOutInMillis;
     }
 
     @Override