+++ mas-foundation/src/main/java/sun/security/x509/EDIPartyName.java
@@ -149,7 +149,7 @@ public class EDIPartyName implements GeneralNameInterface {
     /**
      * Return the assignerName
      *
-     * @returns String assignerName
+     * @return String assignerName
      */
     public String getAssignerName() {
         return assigner;
@@ -158,7 +158,7 @@ public class EDIPartyName implements GeneralNameInterface {
     /**
      * Return the partyName
      *
-     * @returns String partyName
+     * @return String partyName
      */
     public String getPartyName() {
         return party;
@@ -169,7 +169,7 @@ public class EDIPartyName implements GeneralNameInterface {
      * comparison without regard to type of the partyName and
      * the assignerName.
      *
-     * @returns true if the two names match
+     * @return true if the two names match
      */
     public boolean equals(Object other) {
         if (!(other instanceof EDIPartyName))
@@ -229,7 +229,7 @@ public class EDIPartyName implements GeneralNameInterface {
      * certification path verification.
      *
      * @param inputName to be checked for being constrained
-     * @returns constraint type above
+     * @return constraint type above
      * @throws UnsupportedOperationException if name is same type, but comparison operations are
      *          not supported for this name type.
      */
@@ -250,7 +250,7 @@ public class EDIPartyName implements GeneralNameInterface {
      * NameConstraints minimum and maximum bounds and for calculating
      * path lengths in name subtrees.
      *
-     * @returns distance of name from root
+     * @return distance of name from root
      * @throws UnsupportedOperationException if not supported for this name type
      */
     public int subtreeDepth() throws UnsupportedOperationException {