+++ mas-foundation/src/main/java/sun/security/x509/NameConstraintsExtension.java
@@ -411,7 +411,7 @@ implements CertAttrSet<String>, Cloneable {
      * the excluded subtrees state variables.
      *
      * @param cert X509Certificate to be verified
-     * @returns true if certificate verifies successfully
+     * @return true if certificate verifies successfully
      * @throws IOException on error
      */
     public boolean verify(X509Certificate cert) throws IOException {
@@ -489,7 +489,7 @@ implements CertAttrSet<String>, Cloneable {
      * permitted and excluded subtrees variables.
      *
      * @param name GeneralNameInterface name to be verified
-     * @returns true if certificate verifies successfully
+     * @return true if certificate verifies successfully
      * @throws IOException on error
      */
     public boolean verify(GeneralNameInterface name) throws IOException {
@@ -571,7 +571,7 @@ implements CertAttrSet<String>, Cloneable {
      * NameConstraints.
      *
      * @param subject the certificate's subject name
-     * @returns true if certificate verifies successfully
+     * @return true if certificate verifies successfully
      * @throws IOException on error
      */
     public boolean verifyRFC822SpecialCase(X500Name subject) throws IOException {